chore: closes #245 Phase 2 — workspace clippy --fix auto-sweep (v0.5.396)
132 files, +1314/-1648 = -334 LOC net. Mechanical clippy auto-fixes:
- ~43× transmute<i64,f64> → f64::from_bits(i64::cast_unsigned()) (NaN-boxing)
- ~22 unneeded return removals
- ~10 manual div_ceil rewrites
- ~5 manual !RangeInclusive::contains rewrites
- ~5 is_multiple_of rewrites
- ~57 Cell/Vec/Box::new simplifications
- assorted std::mem::take, std::env::var, std::iter::repeat_n usages

Also: cargo fmt --all so the lint gate stays green.

Excluded clippy::not_unsafe_ptr_arg_deref via -A flag — 300 instances
are pub extern "C" fn FFI surface unsafe by contract; marking unsafe
would require the codegen IR to wrap every call site in unsafe {} which
it can't (calls come from LLVM `call`, not Rust). Phase 1 (the strict
gate) needs a project-wide allow directive in clippy.toml first;
deferring to a focused follow-up.

Verified: cargo build --release clean, cargo test --workspace 458/0,
cargo fmt --check clean, gap baseline preserved.

+++ crates/perry-jsruntime/src/bridge.rs
@@ -509,7 +509,7 @@ fn v8_object_to_native(scope: &mut v8::HandleScope<'_>, obj: v8::Local<v8::Objec
         .get_own_property_names(scope, v8::GetPropertyNamesArgs::default())
         .unwrap_or_else(|| v8::Array::new(scope, 0));
 
-    let field_count = names.length() as u32;
+    let field_count = names.length();
 
     // Allocate native object
     let native_obj = unsafe { js_object_alloc(0, field_count) };
